requirements.txt
requests
schedule
PyYAML
psycopg2-binary
